requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
frozenlist==1.4.1
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
python-dotenv
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
magic-filter==1.0.12
moviepy==1.0.3
multidict==6.0.5
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.2
requests==2.31.0
rsa==4.9
tqdm==4.66.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
